chore: Move Test extension to its own file
