+++ src/Uno.UI.RuntimeTests/Tests/Windows_UI_Xaml_Controls/Given_NativeFrame.cs
@@ -12,6 +12,7 @@ using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml;
 using System.Threading.Tasks;
 using System.Diagnostics;
+using Uno.UI.RuntimeTests.Extensions;
 
 namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 {
@@ -82,35 +83,8 @@ namespace Uno.UI.RuntimeTests.Tests.Windows_UI_Xaml_Controls
 			SUT.GoBack();
 
 			await SUT.WaitForPages(1);
-		}		
-	}
-
-	internal static class FrameExtensions
-	{
-		public static int GetAllMyPages(this Frame sut) => sut.EnumerateAllChildren(v => v is MyPage).Count();
-
-		/// Actively waiting for pages to be stacked is
-		/// required as NativeFramePresenter.UpdateStack awaits
-		/// for animations to finish, and there's no way to determine
-		/// from the Frame PoV that the animation is finished.
-		public static async Task WaitForPages(this Frame frame, int count)
-		{
-			var sw = Stopwatch.StartNew();
-
-			while (sw.Elapsed < TimeSpan.FromSeconds(5))
-			{
-				await TestServices.WindowHelper.WaitForIdle();
-
-				if (frame.GetAllMyPages() == count)
-				{
-					break;
-				}
-			}
-
-			Assert.AreEqual(count, frame.GetAllMyPages());
 		}
 	}
-
 	partial class MyPage : Page
 	{
 	}
